[dual-vm] add abi related into types
# Conflicts:
#	Cargo.lock
