+++ types/Cargo.toml
@@ -15,6 +15,8 @@ starcoin-crypto = { workspace = true }
 starcoin-uint = { workspace = true }
 starcoin-vm-types = { workspace = true }
 starcoin-resource-viewer = { workspace = true }
+starcoin-abi-decoder = { workspace = true }
+starcoin-abi-types = { workspace = true }
 thiserror = { workspace = true }
 lazy_static = { workspace = true }
 move-binary-format = { workspace = true }
