+++ types/src/view/decoded_script_view.rs
@@ -1,10 +1,13 @@
 // Copyright (c) The Starcoin Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::view::function_arg_type_view::TypeTagView;
-use crate::view::str_view::StrView;
+use crate::view::{function_arg_type_view::{ModuleIdView, TypeTagView}, str_view::StrView};
+use move_core_types::{account_address::AccountAddress, identifier::Identifier};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+use starcoin_abi_decoder::{
+    DecodedMoveValue, DecodedPackage, DecodedScript, DecodedScriptFunction,
+};
 
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, JsonSchema)]
 pub struct DecodedScriptView {
