+++ types/src/view/mod.rs
@@ -17,7 +17,8 @@ mod block_view;
 mod byte_code_or_script_function;
 mod bytes_view;
 mod chain_info_view;
-// mod decoded_script_view;
+mod code_view;
+mod decoded_script_view;
 mod dry_run_output_view;
 mod event_with_proof_view;
 mod function_arg_type_view;
@@ -25,7 +26,7 @@ mod minted_block_view;
 mod peer_info_view;
 mod raw_block_view;
 mod raw_user_transaction_view;
-// mod resource_view;
+mod resource_view;
 mod scripts_data;
 mod signed_user_transaction_view;
 mod spare_merkle_proof_view;
@@ -41,8 +42,7 @@ mod transaction_info_view;
 mod transaction_output_action;
 mod transaction_output_table_item_action;
 mod transaction_output_view;
-// mod transaction_payload_view;
-mod code_view;
+mod transaction_payload_view;
 mod transaction_request;
 mod transaction_status_view;
 mod transaction_view;
