+++ types/src/view/resource_view.rs
@@ -1,11 +1,11 @@
 // Copyright (c) The Starcoin Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::view::str_view::StrView;
+use crate::view::{function_arg_type_view::StructTagView, str_view::StrView};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+use starcoin_abi_decoder::DecodedMoveValue;
 use std::collections::BTreeMap;
-// use starcoin_abi_decoder::DecodedMoveValue;
 
 #[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
 pub struct ResourceView {
@@ -13,6 +13,16 @@ pub struct ResourceView {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub json: Option<DecodedMoveValue>,
 }
+
+impl From<Vec<u8>> for ResourceView {
+    fn from(v: Vec<u8>) -> Self {
+        Self {
+            raw: StrView(v),
+            json: None,
+        }
+    }
+}
+
 
 #[derive(Default, Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct ListResourceView {
