+++ types/src/view/transaction_event_response.rs
@@ -3,6 +3,8 @@
 
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+use starcoin_abi_decoder::DecodedMoveValue;
+use crate::view::transaction_event_view::TransactionEventView;
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq, JsonSchema)]
 pub struct TransactionEventResponse {
