+++ types/src/view/transaction_payload_view.rs
@@ -1,5 +1,9 @@
+use crate::view::decoded_script_view::{
+    DecodedPackageView, DecodedScriptFunctionView, DecodedScriptView,
+};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+use starcoin_abi_decoder::DecodedTransactionPayload;
 
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, JsonSchema)]
 pub enum TransactionPayloadView {
